Node configuration
Added validateApiKey and getTournament - neither are currently used in appCtrl as of this commit. Only other changes were to remove individual logs and to change from the is_organizer checks that set an api-key to just using the OR setting used in other queries. Passing a null api-key should be sufficient instead of ever using the is_organizer flag.